requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pyswisseph>=2.10.3
geopy>=2.4.1
timezonefinder>=6.5.0
pytz>=2024.2
numpy>=1.26.0
pandas>=2.2.0
plotly>=5.24.0
matplotlib>=3.9.0
wordcloud>=1.9.3
python-multipart>=0.0.12
